+++ tests/KeyboardApiTest.php
@@ -63,7 +63,7 @@ class KeyboardApiTest extends BaseTestCase
             ->getReturnValue();
 
         // checks if the input contains the typed text
-        $this->assertEquals('bar', $value);
+        self::assertEquals('bar', $value);
     }
 
     /**
@@ -80,7 +80,7 @@ class KeyboardApiTest extends BaseTestCase
             ->evaluate('document.activeElement === document.querySelector("#myinput");')
             ->getReturnValue();
 
-        $this->assertFalse($value);
+        self::assertFalse($value);
 
         // press the Tab key
         $page->keyboard()->typeRawKey('Tab');
@@ -90,7 +90,7 @@ class KeyboardApiTest extends BaseTestCase
             ->evaluate('document.activeElement === document.querySelector("#myinput");')
             ->getReturnValue();
 
-        $this->assertTrue($value);
+        self::assertTrue($value);
     }
 
     /**
@@ -128,7 +128,7 @@ class KeyboardApiTest extends BaseTestCase
             ->getReturnValue();
 
         // check if the input contains the typed text twice
-        $this->assertEquals($text.$text, $value);
+        self::assertEquals($text.$text, $value);
     }
 
     /**
@@ -145,7 +145,7 @@ class KeyboardApiTest extends BaseTestCase
             ->press('b')
             ->release();
 
-        $this->assertEquals(0, \count($page->keyboard()->getPressedKeys()));
+        self::assertEquals(0, \count($page->keyboard()->getPressedKeys()));
     }
 
     /**
@@ -167,7 +167,7 @@ class KeyboardApiTest extends BaseTestCase
         $millisecondsElapsed = \round(\hrtime(true) / 1000 / 1000) - $start;
 
         // if this test takes less than 300ms to run (3 keys x 100ms), setKeyInterval is not working
-        $this->assertGreaterThan(300, $millisecondsElapsed);
+        self::assertGreaterThan(300, $millisecondsElapsed);
     }
 
     /**
@@ -191,6 +191,6 @@ class KeyboardApiTest extends BaseTestCase
             ->getReturnValue();
 
         // checks if the input contains the typed text
-        $this->assertSame($text, $value);
+        self::assertSame($text, $value);
     }
 }